Simplify SideDrawer, minimal refactoring to avoid regression
